requirements.txt
langchain>=0.0.350 # for langchain
openai>=1.6.1 # for openai api
python-dotenv>=1.0.0 # for dotenv
google-auth-oauthlib>=1.1.0 # for google auth
google-auth-httplib2>=0.1.1 # for google auth
google-api-python-client>=2.108.0 # for google api
pydantic>=2.5.0 # for pydantic model
rich>=13.7.0 # for rich console
typer>=0.9.0 # for typer cli
loguru>=0.7.2 # for logging